Add jwt check and verification on certain routes

+++ index.js
@@ -3,6 +3,8 @@ const express = require("express");
 const { MongoClient, ObjectId } = require("mongodb");
 require("dotenv").config();
 const cors = require("cors");
+const cookieParser = require("cookie-parser");
+const jwt = require("jsonwebtoken");
 
 // Setup
 const app = express();
@@ -34,27 +36,71 @@ const commentCollection = database.collection("comments");
 const wishlistCollection = database.collection("wishlist");
 const userCollection = database.collection("users");
 
-// Middlewares
-app.use(express.json());
-app.use(
-    cors({
-        origin: "http://localhost:5173",
-        // credentials: true,
-    })
-);
-
 // Utility
 function logger(req, res, next) {
     console.log(`Routing ${req.method} request through ${req.url}`);
     next();
 }
 
+// Middlewares
+app.use(express.json());
+app.use(
+    cors({
+        origin: "http://localhost:5173",
+        credentials: true,
+    })
+);
+app.use(cookieParser());
+const middleman = async (req, res, next) => {
+    const { token } = req.cookies;
+
+    //if client does not send token
+    if (!token) {
+        return res
+            .status(401)
+            .send({ message: "Not Authorized : No Access Token Found" });
+    }
+
+    // Verify token
+    jwt.verify(token, process.env.ACCESS_SECRET, function (err, decoded) {
+        if (err) {
+            return res
+                .status(401)
+                .send({ message: "Not Authorized : Invalid Token" });
+        }
+        // attach decoded user so that others can get it
+        req.user = decoded;
+        next();
+    });
+};
+
 // Routes
+const apiBase = "/api/v1";
 // Home route
 app.get("/", logger, async (req, res) => {
     res.send(`Server is running.......`);
 });
-const apiBase = "/api/v1";
+// JWT route
+app.post(`${apiBase}/get-token`, logger, async (req, res) => {
+    const user = req.body;
+    console.log(`Cookie request user : ${JSON.stringify(user)}`);
+    const secret = process.env.ACCESS_SECRET;
+    const token = jwt.sign(user, secret);
+    res.cookie("token", token, {
+        httpOnly: true,
+        secure: process.env.NODE_ENV === "production",
+        sameSite: process.env.NODE_ENV === "production" ? "none" : "strict",
+    }).send({ success: true });
+});
+app.post(`${apiBase}/delete-token`, logger, async (req, res) => {
+    const user = req.body;
+    console.log(`Cookie delete request user : ${JSON.stringify(user)}`);
+    res.clearCookie("token", {
+        maxAge: 0,
+        secure: process.env.NODE_ENV === "production",
+        sameSite: process.env.NODE_ENV === "production" ? "none" : "strict",
+    }).send({ success: true });
+});
 // Blogs route
 app.get(`${apiBase}/blogs`, logger, async (req, res) => {
     let query = req.query;
@@ -87,20 +133,18 @@ app.get(`${apiBase}/blogs`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.post(`${apiBase}/blogs`, logger, async (req, res) => {
+app.post(`${apiBase}/blogs`, middleman, logger, async (req, res) => {
     const doc = req.body;
     try {
         const result = await blogCollection.insertOne(doc);
         res.send(`Inserted doc at id ${result.insertedId}`);
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.put(`${apiBase}/blogs`, logger, async (req, res) => {
+app.put(`${apiBase}/blogs`, middleman, logger, async (req, res) => {
     const query = req.query;
     try {
         if (query.id) {
@@ -129,7 +173,6 @@ app.put(`${apiBase}/blogs`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
 app.delete(`${apiBase}/blogs`, logger, async (req, res) => {
@@ -143,11 +186,10 @@ app.delete(`${apiBase}/blogs`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
 // Users route
-app.get(`${apiBase}/users`, logger, async (req, res) => {
+app.get(`${apiBase}/users`, middleman, logger, async (req, res) => {
     let query = req.query;
     try {
         // Data soring/filtering based on Query
@@ -155,15 +197,27 @@ app.get(`${apiBase}/users`, logger, async (req, res) => {
             let id = query["id"];
             query = { _id: new ObjectId(id) };
             const result = await userCollection.findOne(query);
-            res.send(result);
-        } else {
-            const cursor = userCollection.find();
-            const result = await cursor.toArray();
-            res.send(result);
+
+            const queryEmail = result.email;
+            const loggedInEmail = req.user.email;
+
+            if (queryEmail !== loggedInEmail) {
+                return res.status(403).send({ message: "Forbidden access" });
+            } else {
+                res.send(result);
+            }
+        }
+        // Commenting out access to all user data. Can be enabled for troubleshoot in production
+        // else {
+        //     const cursor = userCollection.find();
+        //     const result = await cursor.toArray();
+        //     res.send(result);
+        // }
+        else {
+            res.status(404).send({ message: "Not Found" });
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
 app.post(`${apiBase}/users`, logger, async (req, res) => {
@@ -173,49 +227,79 @@ app.post(`${apiBase}/users`, logger, async (req, res) => {
         res.send(`Inserted doc at id ${result.insertedId}`);
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.put(`${apiBase}/users`, logger, async (req, res) => {
+app.put(`${apiBase}/users`, middleman, logger, async (req, res) => {
     const query = req.query;
     try {
         if (query.id) {
             let id = query["id"];
             const filter = { _id: new ObjectId(id) };
-            const options = { upsert: true };
-            const updatedData = req.body;
-            const newData = {
-                $set: {
-                    name: updatedData.name,
-                    image_url: updatedData.image_url,
-                    email: updatedData.email,
-                    date_registered: updatedData.date_registered,
-                },
-            };
-            const result = await userCollection.updateOne(
-                filter,
-                newData,
-                options
-            );
-            res.send(result);
+            const user = await userCollection.findOne(filter);
+            if (user.email) {
+                const queryEmail = user.email;
+                const loggedInEmail = req.user.email;
+                if (queryEmail !== loggedInEmail) {
+                    return res
+                        .status(403)
+                        .send({ message: "Forbidden access" });
+                } else {
+                    const options = { upsert: true };
+                    const updatedData = req.body;
+                    const newData = {
+                        $set: {
+                            name: updatedData.name,
+                            image_url: updatedData.image_url,
+                            email: updatedData.email,
+                            date_registered: updatedData.date_registered,
+                        },
+                    };
+                    const result = await userCollection.updateOne(
+                        filter,
+                        newData,
+                        options
+                    );
+                    res.send(result);
+                }
+            } else {
+                res.status(404).send({
+                    message: "Not Found",
+                });
+            }
+        } else {
+            res.send({ message: "invalid" });
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.delete(`${apiBase}/users`, logger, async (req, res) => {
+app.delete(`${apiBase}/users`, middleman, logger, async (req, res) => {
     try {
-        const query = req.query;
         if (query.id) {
             let id = query["id"];
             const filter = { _id: new ObjectId(id) };
-            const result = await userCollection.deleteOne(filter);
-            res.send(result);
+            const user = await userCollection.findOne(filter);
+            if (user.email) {
+                const queryEmail = user.email;
+                const loggedInEmail = req.user.email;
+                if (queryEmail !== loggedInEmail) {
+                    return res
+                        .status(403)
+                        .send({ message: "Forbidden access" });
+                } else {
+                    const result = await userCollection.deleteOne(filter);
+                    res.send(result);
+                }
+            } else {
+                res.status(404).send({
+                    message: "Not Found",
+                });
+            }
+        } else {
+            res.send({ message: "invalid" });
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
 // Comments route
@@ -250,20 +334,18 @@ app.get(`${apiBase}/comments`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.post(`${apiBase}/comments`, logger, async (req, res) => {
+app.post(`${apiBase}/comments`, middleman, logger, async (req, res) => {
     const doc = req.body;
     try {
         const result = await commentCollection.insertOne(doc);
         res.send(`Inserted doc at id ${result.insertedId}`);
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.put(`${apiBase}/comments`, logger, async (req, res) => {
+app.put(`${apiBase}/comments`, middleman, logger, async (req, res) => {
     const query = req.query;
     try {
         if (query.id) {
@@ -287,10 +369,9 @@ app.put(`${apiBase}/comments`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.delete(`${apiBase}/comments`, logger, async (req, res) => {
+app.delete(`${apiBase}/comments`, middleman, logger, async (req, res) => {
     try {
         const query = req.query;
         if (query.id) {
@@ -301,11 +382,10 @@ app.delete(`${apiBase}/comments`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
 // Wishlist route
-app.get(`${apiBase}/wishlist`, logger, async (req, res) => {
+app.get(`${apiBase}/wishlist`, middleman, logger, async (req, res) => {
     let query = req.query;
     try {
         const sortFiled = {};
@@ -317,34 +397,45 @@ app.get(`${apiBase}/wishlist`, logger, async (req, res) => {
             const result = await wishlistCollection.findOne(query);
             res.send(result);
         } else if (query.owner) {
-            filter.owner = query.owner;
-            if (query.sort) {
-                if (query.sort.startsWith("-")) {
-                    sortFiled[query.sort.slice(1)] = -1;
-                } else {
-                    sortFiled[query.sort] = 1;
+            let id = query.owner;
+            query = { _id: new ObjectId(id) };
+            const result = await userCollection.findOne(query);
+
+            const queryEmail = result.email;
+            const loggedInEmail = req.user.email;
+
+            if (queryEmail !== loggedInEmail) {
+                return res.status(403).send({ message: "Forbidden access" });
+            } else {
+                filter.owner = id;
+                if (query.sort) {
+                    if (query.sort.startsWith("-")) {
+                        sortFiled[query.sort.slice(1)] = -1;
+                    } else {
+                        sortFiled[query.sort] = 1;
+                    }
                 }
+                const cursor = wishlistCollection.find(filter).sort(sortFiled);
+                const result = await cursor.toArray();
+                res.send(result);
             }
-            const cursor = wishlistCollection.find(filter).sort(sortFiled);
-            const result = await cursor.toArray();
-            res.send(result);
+        } else {
+            res.status(404).send({ message: "Not Found" });
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.post(`${apiBase}/wishlist`, logger, async (req, res) => {
+app.post(`${apiBase}/wishlist`, middleman, logger, async (req, res) => {
     const doc = req.body;
     try {
         const result = await wishlistCollection.insertOne(doc);
         res.send(`Inserted doc at id ${result.insertedId}`);
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
-app.delete(`${apiBase}/wishlist`, logger, async (req, res) => {
+app.delete(`${apiBase}/wishlist`, middleman, logger, async (req, res) => {
     try {
         const query = req.query;
         if (query.id) {
@@ -355,6 +446,5 @@ app.delete(`${apiBase}/wishlist`, logger, async (req, res) => {
         }
     } catch (error) {
         console.log(`Error while routing ${req.url} : ${error}`);
-        res.send(`{Erorr : ${error} }`);
     }
 });
